+++ CHANGELOG.rst
@@ -10,7 +10,7 @@ Changelog
 - Write attributes via dictionary
 - Macros to check file and dataset validity
 - Improved handling of strings in numpy array in attributes
-
+- Can convert single-entry numpy array numeric attributes to Python native types
 
 0.1.0 (18-05-04)
 ----------------